requirements.txt
cairosvg==2.7.0
Pillow
svgpathtools
